refacto modules for indicators
