+++ app/add_ons/indicators/mma.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pyqtgraph as pg
 
-from libs.indicators_widget import Indicator, InputField, ChoiceField
+from utils.indicators_utils import Indicator, InputField, ChoiceField
 
 
 class MMA(Indicator):
